requirements.txt
apprise==1.0.0
cachetools==4.2.2
fire==0.4.0
googletrans==3.1.0a0
httpx==0.13.3
newsapi==0.1.1
newsapi_python==0.2.6
numpy==1.22.4
openai==0.23.0
pandas==1.4.3
python_dateutil==2.8.2
PyYAML==6.0
schedule==1.1.0
tqdm==4.64.0
tweepy==4.10.1
